requirements.txt
affine==2.4.0
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
cfgrib==0.9.10.4
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
dask==2024.2.0
"dask[diagnostics]"==2024.2.0
eccodes==1.6.1
ecmwf-api-client==1.6.3
findlibs==0.0.5
fiona==1.9.5
fsspec==2024.2.0
geopandas==0.14.3
importlib-metadata==7.0.1
jinja2==3.1.4
locket==1.0.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
partd==1.4.1
pycparser==2.21
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
rasterio==1.3.9
rioxarray==0.15.1
scipy==1.14.0
shapely==2.0.2
six==1.16.0
snuggs==1.4.7
toolz==0.12.1
tzdata==2024.1
xarray==2024.1.1
zipp==3.17.0
